fix(minesweeper): fix x and y confusion in field_matrix
Field matrix is a vector of vectors. Each row of the field
is a vector and each element in that vector is a cell. So,
to access an element in the matrix you have to give it
the y coordinate before the x, like this: field_matrix[row][cell].
I was doing it the other way around before, now it is fixed.